+++ crates/helpers/Cargo.toml
@@ -4,5 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-aidoku_imports = { path = "../imports" }
 aidoku_macros = { path = "../macros" }
